Rename EmojiCoverter to Html::ParseEmoji
